BUG: TDA.__floordiv__ with NaT (#35583)
* BUG: TDA.__floordiv__ with NaT

* whatsnew

+++ pandas/core/arrays/timedeltas.py
@@ -628,7 +628,7 @@ class TimedeltaArray(dtl.DatetimeLikeArrayMixin, dtl.TimelikeOps):
             result = self.asi8 // other.asi8
             mask = self._isnan | other._isnan
             if mask.any():
-                result = result.astype(np.int64)
+                result = result.astype(np.float64)
                 result[mask] = np.nan
             return result
